Move todo data to the component props
